requirements.txt
absl-py==2.1.0
asgiref==3.8.1
attrs==24.2.0
azure-common==1.1.28
azure-core==1.31.0
azure-storage-blob==12.0.0
azure-storage-common==2.1.0
boto3==1.35.26
botocore==1.35.26
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.3.0
cryptography==43.0.1
cycler==0.12.1
decorator==4.4.2
Django==5.1.1
django-storages==1.14.4
flatbuffers==24.3.25
fonttools==4.53.1
idna==3.10
imageio==2.35.1
imageio-ffmpeg==0.5.1
isodate==0.6.1
jax==0.4.31
jaxlib==0.4.31
jmespath==1.0.1
kiwisolver==1.4.7
matplotlib==3.9.2
mediapipe==0.10.14
ml-dtypes==0.4.0
moviepy==1.0.3
msrest==0.7.1
numpy==1.26.4
oauthlib==3.2.2
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opt-einsum==3.3.0
packaging==24.1
pillow==10.4.0
proglog==0.1.10
protobuf==4.25.4
pycparser==2.22
pydub==0.25.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
requests==2.32.3
requests-oauthlib==2.0.0
s3transfer==0.10.2
scipy==1.14.1
six==1.16.0
sounddevice==0.5.0
SpeechRecognition==3.10.4
sqlparse==0.5.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
whitenoise==6.7.0
torch==2.4.0
torchaudio==2.4.0
torchvision==0.19.0
transformers==4.45.1
